+++ packages/amplify-codegen-e2e-core/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.1](https://github.com/aws-amplify/amplify-codegen/compare/@aws-amplify/amplify-codegen-e2e-core@1.6.0...@aws-amplify/amplify-codegen-e2e-core@1.6.1) (2023-09-26)
+
+**Note:** Version bump only for package @aws-amplify/amplify-codegen-e2e-core
+
 # [1.6.0](https://github.com/aws-amplify/amplify-codegen/compare/@aws-amplify/amplify-codegen-e2e-core@1.5.4...@aws-amplify/amplify-codegen-e2e-core@1.6.0) (2023-09-19)
 
 ### Features